pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pdf2anki"
version = "0.1.0"
description = "A CLI tool for converting PDFs into Anki flashcards."
readme = "README.md"
license = {text = "Custom Personal Use License v1.4"}
requires-python = "~=3.11.0"
authors = [
    {name = "Martin Krause", email = "martinkrausemedia@gmail.com"}
]
urls = { "Homepage" = "https://your-repo-url/pdf2anki" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: Other/Proprietary License",
    "Operating System :: OS Independent",
    "Topic :: Utilities",
    "Intended Audience :: Education"
]
dependencies = [
    "ollama>=0.3.3",
    "pdf2image>=1.16.3",
    "Pillow>=9.1.0"
]

[project.scripts]
pdf2anki = "pdf2anki.core:cli_invoke"
